Update code structure and add more functionality
The structure of the code has been altered to better
resemble Facebook's own example bot. The use of the function
'callSendApi' is cleaner and will allow better extension
into creating messages that aren't just text.

RetrieveUserInfo is a function to add a more personal touch
to Miranda. However, due to async programming, this was a lot
harder than I anticipated. I used an example from someone else
that involved utilizing callbacks, however I'm still not entirely
familiar with the concept and need to keep investigating.

Finally, changed how Miranda identified typical greetings by
using regex. This is extremely dumb and will be changed in the
future.